requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
httpx>=0.27.0
protobuf>=4.25.0
ray[default]>=2.31.0
torch>=2.3.0
transformer-engine>=1.6.0
deepspeed>=0.14.0
qdrant-client>=1.9.0
pydantic<2.0
ruff>=0.5.0
